emacs: JSON based search improvements.
Fix variable name typos.

Attempt to avoid display flashing by removing newlines from the
inserted string.

Ease debugging by moving the `condition-case' closer into the code
that it is really intended to cover (the parsing of the JSON object),
allowing errors in the remaining code to still fire correctly. As a
consequence, test for :json-eof when examining the next character to
be parsed, as that throws an error that was previously covered by the
`condition-case'.